refacotry

+++ src/Localidade/Dominio/Dominio.Localidade/Extensions/MesorregiaoExtension.cs
@@ -1,9 +1,4 @@
 ﻿using Dominio.Localidade.Enum;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Dominio.Localidade.Extensions
 {
@@ -12,14 +7,14 @@ namespace Dominio.Localidade.Extensions
         public static Regiao Regiao(this Mesorregiao mesorregiao)
         {
             var value = (int)mesorregiao;
-            var result = (Regiao)((int)value / 10000);
+            var result = (Regiao)(value / 10000);
             return result;
         }
 
         public static Estado Estado(this Mesorregiao mesorregiao)
         {
             var value = (int)mesorregiao;
-            var result = (Estado)((int)value / 1000);
+            var result = (Estado)(value / 1000);
             return result;
         }
     }
